requirements.txt
flask
google-generativeai
guardrails-ai
